+++ components/base/PrimaryButton.vue
@@ -1,5 +1,5 @@
 <template>
-  <button class="font-bebas text-xl flex p-4 gap-5 bg-black text-white dark:bg-white dark:text-black hover:bg-[#714AFF] dark:hover:text-white transition-all" @click="handleClick">
+  <button class="font-bebas text-xl flex p-4 gap-5 bg-black text-white dark:bg-white dark:text-black hover:bg-[#714AFF] dark:hover:text-white transition-all duration-500" @click="handleClick">
     {{ buttonName }}
   </button>
 </template>
@@ -10,11 +10,17 @@ export default {
     buttonName: {
       type: String,
       default: "PRIMARY BUTTON"
+    },
+    link: {
+      type: String,
+      default: "" 
     }
   },
   methods: {
     handleClick() {
-      if(this.onClick) {
+      if (this.link) {
+        window.open(this.link, "_blank")
+      } else if (this.onClick) {
         this.onClick()
       }
     }