+++ components/base/SpecialButton.vue
@@ -1,5 +1,5 @@
 <template>
-  <button class="font-bebas text-xl flex p-4 px-8 gap-5 rounded-full bg-black text-white dark:bg-white dark:text-black hover:bg-[#714AFF] dark:hover:text-white transition-all duration-500" @click="handleClick">
+  <button class="font-bebas text-2xl flex p-5 px-8 gap-5 rounded-full bg-white text-[#714AFF] hover:bg-[#714AFF] hover:text-white transition-all duration-500" @click="handleClick">
     <component :is="icon" class="w-6 h-6" />
     {{ buttonName }}
   </button>
@@ -10,7 +10,7 @@ export default {
   props: {
     buttonName: {
       type: String,
-      default: "SECONDARY BUTTON"
+      default: "SPECIAL BOSS"
     },
     icon: {
       type: Object